SHA1: 64-bit system fix
Same fix as applied to ezQuake
https://github.com/ezQuake/ezquake-source/commit/975211da3ec23a54220c353f89657fd40f416854

+++ src/sha1.h
@@ -19,13 +19,13 @@ A million repetitions of "a"
 /* #define SHA1HANDSOFF * Copies data before messing with it. */
 
 typedef struct {
-    unsigned long state[5];
-    unsigned long count[2];
+    unsigned int state[5];
+    unsigned int count[2];
     unsigned char buffer[64];
 } SHA1_CTX;
 
 #define DIGEST_SIZE 20
-//void SHA1Transform(unsigned long state[5], unsigned char buffer[64]);
+//void SHA1Transform(unsigned int state[5], unsigned char buffer[64]);
 //void (SHA1_CTX* context);
 //void (SHA1_CTX* context, unsigned char* data, unsigned int len);
 //void (unsigned char digest[DIGEST_SIZE], SHA1_CTX* context);
